Add support for primary constructor assignments as well
