+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/FindSymbols/LinkedFileReferenceLocationEqualityComparer.cs
