+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/VisualBasic/LanguageServices/InitializeParameter/VisualBasicInitializeParameterService.vb
@@ -3,6 +3,7 @@
 ' See the LICENSE file in the project root for more information.
 
 Imports System.Composition
+Imports System.Threading
 Imports Microsoft.CodeAnalysis.Editing
 Imports Microsoft.CodeAnalysis.Host.Mef
 Imports Microsoft.CodeAnalysis.InitializeParameter
@@ -19,11 +20,15 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.InitializeParameter
         Public Sub New()
         End Sub
 
+        Protected Overrides Function IsFunctionDeclaration(node As SyntaxNode) As Boolean
+            Return InitializeParameterHelpers.IsFunctionDeclaration(node)
+        End Function
+
         Protected Overrides Sub InsertStatement(editor As SyntaxEditor, functionDeclaration As SyntaxNode, returnsVoid As Boolean, statementToAddAfter As SyntaxNode, statement As StatementSyntax)
             InitializeParameterHelpers.InsertStatement(editor, functionDeclaration, statementToAddAfter, statement)
         End Sub
 
-        Public Overrides Function GetBody(methodNode As SyntaxNode) As SyntaxNode
+        Protected Overrides Function GetBody(methodNode As SyntaxNode) As SyntaxNode
             Return InitializeParameterHelpers.GetBody(methodNode)
         End Function
 
@@ -34,6 +39,11 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.InitializeParameter
         Protected Overrides Function TryUpdateTupleAssignment(blockStatement As IBlockOperation, parameter As IParameterSymbol, fieldOrProperty As ISymbol, editor As SyntaxEditor) As Boolean
             ' Not supported in VB
             Return False
+        End Function
+
+        Protected Overrides Function TryAddAssignmentForPrimaryConstructorAsync(document As Document, parameter As IParameterSymbol, fieldOrProperty As ISymbol, cancellationToken As CancellationToken) As Task(Of Solution)
+            ' Nothing to do in VB.
+            Return Task.FromResult(document.Project.Solution)
         End Function
     End Class
 End Namespace
